# api_gateway/OpenAPI.yml
openapi: 3.0.3
info:
  title: StudyFlow API
  version: 1.0.0

  security:
    - telegramAuth: []
components:
  securitySchemes:
    telegramAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Two formats supported:

        **1. Telegram Mini App initData (production):**
        `tma <initData from Telegram.WebApp.initData>`
        - initData is validated per Telegram Bot API spec (HMAC-SHA256 with bot token)
        - No client-side secret needed
        - TTL: 24 hours from auth_date

        **2. Legacy HMAC (development only):**
        `telegram <telegramId>:<utcTimestamp>:<hmacHex>`
        - HMAC-SHA256 of `{telegramId}:{utcTimestamp}` with TELEGRAM_SECRET as key
        - ±5 minute clock skew
        - Disable in production with AUTH_DISABLE_LEGACY_HMAC=true
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      example:
        error: Not Found
...
    Assignment:
      type: object
      properties:
        id:
          type: string
        tutorId:
          type: string
        studentId:
          type: string
        title:
          type: string
        description:
          type: string
        fileId:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    Submission:
      type: object
      properties:
        id:
          type: string
        assignmentId:
          type: string
        fileId:
          type: string
        comment:
          type: string
        submittedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    Feedback:
      type: object
      properties:
        id:
          type: string
        submissionId:
          type: string
        fileId:
          type: string
        comment:
          type: string
        grade:
          type: integer
          minimum: 1
          maximum: 5
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    HomeworkFileURL:
      type: object
      properties:
        url:
          type: string
    AssignmentStatus:
      type: string
      enum:
        - UNSENT
        - UNREVIEWED
        - REVIEWED
        - OVERDUE
...
paths:
...
  /homework/assignments:
    post:
      summary: Create assignment
      operationId: createAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tutorId:
                  type: string
                studentId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                fileId:
                  type: string
                dueDate:
                  type: string
                  format: date-time
              required:
                - tutor_id
                - student_id
                - title
                - description
                - file_id
      responses:
        '200':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Student not found or invalid relationship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List assignments by tutor, student, or pair
      operationId: listAssignments
      parameters:
        - name: tutor_id
          in: query
          schema:
            type: string
        - name: student_id
          in: query
          schema:
            type: string
        - name: status_filter
          in: query
          schema:
            $ref: '#/components/schemas/AssignmentStatus'
          style: form
          explode: true
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/assignments/{id}:
    patch:
      summary: Update assignment
      operationId: updateAssignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                fileId:
                  type: string
                dueDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete assignment
      operationId: deleteAssignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Assignment deleted
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/assignments/{assignment_id}/file-url:
    get:
      summary: Get assignment file
      operationId: getAssignmentFile
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeworkFileURL'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/assignments/{assignment_id}/submissions:
    get:
      summary: List submissions by assignment
      operationId: listSubmissionsByAssignment
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Submission'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/assignments/{assignment_id}/feedbacks:
    get:
      summary: List feedbacks by assignment
      operationId: listFeedbacksByAssignment
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of feedbacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/submissions:
    post:
      summary: Create submission
      operationId: createSubmission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assignmentId:
                  type: string
                fileId:
                  type: string
                comment:
                  type: string
              required:
                - assignment_id
                - file_id
      responses:
        '200':
          description: Submission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/submissions/{submission_id}/file-url:
    get:
      summary: Get submission file
      operationId: getSubmissionFile
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeworkFileURL'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/feedbacks:
    post:
      summary: Create feedback
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submissionId:
                  type: string
                fileId:
                  type: string
                comment:
                  type: string
                grade:
                  type: integer
                  minimum: 1
                  maximum: 5
              required:
                - submission_id
                - file_id
      responses:
        '200':
          description: Feedback created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/feedbacks/{id}:
    patch:
      summary: Update feedback
      operationId: updateFeedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fileId:
                  type: string
                comment:
                  type: string
                grade:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '200':
          description: Feedback updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Feedback not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /homework/feedbacks/{feedback_id}/file-url:
    get:
      summary: Get feedback file
      operationId: getFeedbackFile
      parameters:
        - name: feedback_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeworkFileURL'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'